+++ CHANGELOG.md
@@ -9,7 +9,7 @@ Features that are in development, but are not released yet. Does not include stu
 
 ### Added
 - [Confirmation before executing a shell command (#35)](https://github.com/Taitava/obsidian-shellcommands/issues/35)
-- [Settings: Execute now icon button for each command](https://github.com/Taitava/obsidian-shellcommands/issues/30)
+- [Settings: Execute now icon button for each command (#30)](https://github.com/Taitava/obsidian-shellcommands/issues/30)
 
 ### Changed
 - [Working directory: support a relative path (#28)](https://github.com/Taitava/obsidian-shellcommands/issues/28)